create logic to prepare new account thymeleaf rendering(call prepareNewAccount service)
This is to display the correct information after we press "Create New Bank Account" button.
The correct information: Account #(the next account number available for that user based on the number of accounts they have).
If they have 2 accounts -> Account #3 is displayed since it would be creating the Users' THIRD account if submitted.